#1: Initial CD implementation
./microCI.py -d/--deployments FILE -i/--immediate

For each deployment in deployments file, check branch in repository every <interval> minutes. If new commits found, execute action and create issue in repository if action fails. When 'immediate' flag is specified, action will be executed for every commit including the current, otherwise, for every commit after the current one. Token specified in deployment is used to access repository and to create issues.